+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added public keyword to binding consts
+
 ## v0.11.6 (2024-08-21)
 * Made class constants PascalCase